+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2
+
+- feature: add Ability command to make ability checks
+- feature: add Save command to make saving throws
+- feature: add flavor as an option
+
 # 0.1.1
 
 - fill in module description